requirements.txt
uvicorn
python-dotenv
fastapi
sqlalchemy
asyncpg
requests